model_training/requirements.txt
torch
torchvision
albumentations
albumentations.pytorch
pillow
tqdm
numpy
matplotlib